Remove `SocketAddr`'s `From` conversions for owned and borrowed socket types.
It occurred to me that, although these make sense for use in a child process that will be inheriting some sockets, they do *not* make sense in a *parent* process that's preparing some sockets to be inherited.

Moreover, these conversions can only instantiate the `Inherit` variant, not the `InheritStdin` or `SystemdNumeric` variant.